deploy/requirements.txt
attrs==25.4.0
aws-cdk-lib==2.215.0
aws-cdk.asset-awscli-v1==2.2.242
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==48.20.0
cattrs==25.3.0
constructs==10.4.3
importlib_resources==6.5.2
jsii==1.119.0
publication==0.0.3
python-dateutil==2.9.0.post0
six==1.17.0
typeguard==4.2.1
typing_extensions==4.15.0
